Create Model flow changes

+++ app/src/main/java/com/aiavatar/app/core/domain/model/CreateModelData.kt
@@ -1,7 +1,7 @@
 package com.aiavatar.app.core.domain.model
 
 data class CreateModelData(
-    val statusId: Int,
+    val statusId: Long,
     val guestUserId: String,
     val eta: Long
 )
